change exported in manifest and added simple array to test
