+++ app/src/main/AndroidManifest.xml
@@ -17,13 +17,13 @@
         tools:targetApi="31">
         <activity
             android:name=".activity.ParticipantListActivity"
-            android:exported="false" />
+            android:exported="true" />
         <activity
             android:name=".activity.LoginPageActivity"
-            android:exported="false" />
+            android:exported="true" />
         <activity
             android:name=".activity.ModeratorActivity"
-            android:exported="false" />
+            android:exported="true" />
         <activity
             android:name=".MainActivity"
             android:exported="true">
