+++ app/src/main/java/com/group7/meetr/activity/ParticipantListActivity.java
@@ -1,42 +1,40 @@
 package com.group7.meetr.activity;
 
+import android.content.Intent;
+import android.os.Bundle;
+
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import android.os.Bundle;
-import android.util.Log;
-
 import com.google.firebase.database.FirebaseDatabase;
+import com.group7.meetr.MainActivity;
 import com.group7.meetr.R;
 import com.group7.meetr.data.remote.FirebaseRealtimeDatabase;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
 public class ParticipantListActivity extends AppCompatActivity {
 
+     RecyclerView participantListRecyclerView;
+    ParticipantListAdapter adapter;
+    FirebaseRealtimeDatabase firebaseRealtimeDatabase;
+    String[] emails;
 
-        @Override
-        protected void onCreate(Bundle savedInstanceState) {
-            super.onCreate(savedInstanceState);
-            setContentView(R.layout.activity_participant_list);
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_participant_list);
+        firebaseRealtimeDatabase = new FirebaseRealtimeDatabase(FirebaseDatabase.getInstance(), "meetingID", adapter);
+        firebaseRealtimeDatabase.addParticipantsListener();
 
-            RecyclerView recyclerView = findViewById(R.id.ParticipantList);
+        emails = new String[]{};
+        participantListRecyclerView = findViewById(R.id.ParticipantList);
+        participantListRecyclerView.setLayoutManager(new LinearLayoutManager(this));
+        adapter = new ParticipantListAdapter(this, adapter.getParticipants());
+        participantListRecyclerView.setAdapter(adapter);
 
-            //TODO: Be firebase om listan av namn. Alla namn över ska bort.
+        Intent intent = new Intent(ParticipantListActivity.this, MainActivity.class);
+        startActivity(intent);
 
-            ParticipantListAdapter adapter = new ParticipantListAdapter();
-            //String[] names = myAdapter.getParticipants();
-            Log.d("TAG", "onCreate: names" + Arrays.toString(adapter.getParticipants()));
-            FirebaseDatabase database = FirebaseDatabase.getInstance("https://meetr-android-default-rtdb.europe-west1.firebasedatabase.app/");
-            //TODO: Move this code to a more appropriate place.
-            FirebaseRealtimeDatabase realtimeDatabase = new FirebaseRealtimeDatabase(database, "7", adapter);
-            realtimeDatabase.addParticipantsListener();
 
-            recyclerView.setLayoutManager(new LinearLayoutManager(this));
-            recyclerView.setAdapter(new ParticipantListAdapter(getApplicationContext(), adapter.getParticipants()));
-
-        }
     }
+}
